# src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Snapshot x402 API",
    "version": "1.0.0",
    "summary": "Pay-per-request Snapshot governance data for agents, paid in USDC on Algorand via x402.",
...
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This deployment"
    }
  ],
...
  "paths": {
...
    "/v1/spaces/{id}": {
      "get": {
        "operationId": "getSpace",
        "summary": "Space profile, settings and activity stats",
        "description": "One Snapshot space: identity, voting strategies (names and symbols), voting settings, admins, treasuries, delegation portal and activity counters.\n\nDefault price: 0.01 USDC per successful request (x402 exact scheme). Failed requests (4xx/5xx) are not settled.",
        "tags": ["Governance data"],
        "security": [
          {
            "x402": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Snapshot space id (ENS-style name, case-insensitive, max 64 chars), e.g. ens.eth, aavedao.eth.",
            "schema": {
              "$ref": "#/components/schemas/SpaceId"
            },
            "example": "ens.eth"
          }
        ],
        "x-x402": {
          "price": "0.01",
          "currency": "USDC",
          "scheme": "exact",
          "network": "algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73ktiC1qzkkit8="
        },
        "responses": {
          "200": {
            "description": "Paid response. The PAYMENT-RESPONSE header carries the settlement receipt.",
            "headers": {
              "PAYMENT-RESPONSE": {
                "$ref": "#/components/headers/PaymentResponse"
              },
              "Cache-Control": {
                "description": "private, max-age=300",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Space"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "404": {
            "description": "Space not found",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "Space not found"
                  }
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "421": {
            "$ref": "#/components/responses/Misdirected"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/UpstreamTimeout"
          }
        }
      }
    },
    "/v1/spaces/{id}/proposals": {
      "get": {
        "operationId": "listSpaceProposals",
        "summary": "Proposals of a space, newest first",
        "description": "Proposals of one space ordered by creation time (newest first), with live or final scores. Body text is not included; use GET /v1/proposals/{id}?include=body.\n\nDefault price: 0.01 USDC per successful request (x402 exact scheme). Failed requests (4xx/5xx) are not settled.",
        "tags": ["Governance data"],
        "security": [
          {
            "x402": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Snapshot space id (ENS-style name, case-insensitive, max 64 chars), e.g. ens.eth, aavedao.eth.",
            "schema": {
              "$ref": "#/components/schemas/SpaceId"
            },
            "example": "ens.eth"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "description": "Filter by state. Omit for all states.",
            "schema": {
              "type": "string",
              "enum": ["active", "pending", "closed"]
            }
          },
          {
            "name": "first",
            "in": "query",
            "required": false,
            "description": "Page size.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "description": "Offset for pagination.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 5000,
              "default": 0
            }
          }
        ],
        "x-x402": {
          "price": "0.01",
          "currency": "USDC",
          "scheme": "exact",
          "network": "algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73ktiC1qzkkit8="
        },
        "responses": {
          "200": {
            "description": "Paid response. The PAYMENT-RESPONSE header carries the settlement receipt.",
            "headers": {
              "PAYMENT-RESPONSE": {
                "$ref": "#/components/headers/PaymentResponse"
              },
              "Cache-Control": {
                "description": "private, max-age=60 (30 when state=active)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "404": {
            "description": "Space not found",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "Space not found"
                  }
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "421": {
            "$ref": "#/components/responses/Misdirected"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/UpstreamTimeout"
          }
        }
      }
    },
    "/v1/proposals/{id}": {
      "get": {
        "operationId": "getProposal",
        "summary": "Proposal with results",
        "description": "One proposal with per-choice results (score, share of total, score per strategy), quorum, strategies and state. Add include=body for the Markdown body.\n\nDefault price: 0.01 USDC per successful request (x402 exact scheme). Failed requests (4xx/5xx) are not settled.",
        "tags": ["Governance data"],
        "security": [
          {
            "x402": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Snapshot proposal id: 0x + 64 hex (current), or a legacy IPFS CID (Qm... or baf...).",
            "schema": {
              "$ref": "#/components/schemas/ProposalId"
            },
            "example": "0xcd8d33f6531ed3baa3f728ca8509c66ce7505877b349e6127981e8254adf9a22"
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "description": "Set to body to include the proposal body (Markdown).",
            "schema": {
              "type": "string",
              "enum": ["body"]
            }
          }
        ],
        "x-x402": {
          "price": "0.01",
          "currency": "USDC",
          "scheme": "exact",
          "network": "algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73ktiC1qzkkit8="
        },
        "responses": {
          "200": {
            "description": "Paid response. The PAYMENT-RESPONSE header carries the settlement receipt.",
            "headers": {
              "PAYMENT-RESPONSE": {
                "$ref": "#/components/headers/PaymentResponse"
              },
              "Cache-Control": {
                "description": "private, max-age=3600 when closed and scores final; 30 when active; 60 otherwise",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Proposal"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "404": {
            "description": "Proposal not found",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "Proposal not found"
                  }
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "421": {
            "$ref": "#/components/responses/Misdirected"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/UpstreamTimeout"
          }
        }
      }
    },
    "/v1/proposals/{id}/votes": {
      "get": {
        "operationId": "listProposalVotes",
        "summary": "Votes on a proposal",
        "description": "Votes on one proposal, ordered by voting power (default) or time, with the proposal choices so choice indexes can be mapped to labels. Choice format depends on the proposal type: single-choice and basic use a 1-based index; approval, ranked-choice and copeland use an array of 1-based indexes; weighted and quadratic use an object of index to weight. For shutter-private proposals the choice is an encrypted hex string until the proposal closes.\n\nDefault price: 0.01 USDC per successful request (x402 exact scheme). Failed requests (4xx/5xx) are not settled.",
        "tags": ["Governance data"],
        "security": [
          {
            "x402": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Snapshot proposal id: 0x + 64 hex (current), or a legacy IPFS CID (Qm... or baf...).",
            "schema": {
              "$ref": "#/components/schemas/ProposalId"
            },
            "example": "0xcd8d33f6531ed3baa3f728ca8509c66ce7505877b349e6127981e8254adf9a22"
          },
          {
            "name": "first",
            "in": "query",
            "required": false,
            "description": "Page size.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "description": "Offset for pagination.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 5000,
              "default": 0
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "required": false,
            "description": "Sort key, always descending.",
            "schema": {
              "type": "string",
              "enum": ["vp", "created"],
              "default": "vp"
            }
          }
        ],
        "x-x402": {
          "price": "0.01",
          "currency": "USDC",
          "scheme": "exact",
          "network": "algorand:wGHE2Pwdvd7S12BL5FaOP20EGYesN73ktiC1qzkkit8="
        },
        "responses": {
          "200": {
            "description": "Paid response. The PAYMENT-RESPONSE header carries the settlement receipt.",
            "headers": {
              "PAYMENT-RESPONSE": {
                "$ref": "#/components/headers/PaymentResponse"
              },
              "Cache-Control": {
                "description": "private, max-age=3600 when proposal closed and scores final; 30 otherwise",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "404": {
            "description": "Proposal not found",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "Proposal not found"
                  }
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "421": {
            "$ref": "#/components/responses/Misdirected"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "504": {
            "$ref": "#/components/responses/UpstreamTimeout"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "PAYMENT-SIGNATURE",
        "description": "x402 v2 payment: base64-encoded JSON PaymentPayload built from one entry of the PAYMENT-REQUIRED header returned by a 402. See https://x402.org."
      }
    },
    "headers": {
...
      "PaymentResponse": {
        "description": "Base64 JSON SettleResponse (x402 v2).",
        "schema": {
          "type": "string",
          "contentEncoding": "base64",
          "contentMediaType": "application/json",
          "contentSchema": {
            "$ref": "#/components/schemas/SettleResponse"
          }
        }
      },
      "CacheControlNoStore": {
        "description": "no-store",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "BadGateway": {
        "description": "Snapshot upstream failed (upstream_error), or the payment facilitator failed or timed out (facilitator_error). Not charged, except that a facilitator timeout during settlement can still be confirmed on chain.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid input. Checked after the payment is verified; the payment is not settled, so it is not charged.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "bad_request",
                "message": "first must be an integer from 1 to 100"
              }
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "Only GET is allowed on paid routes (method_not_allowed). Not charged.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Misdirected": {
        "description": "The request used a host other than the public URL of this deployment (misdirected_request); the message names the URL to use. Not charged.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PaymentRequired": {
        "description": "Payment required. Decode the PAYMENT-REQUIRED header, pay one of accepts[], retry with PAYMENT-SIGNATURE. Also returned when a payment is rejected (error.code payment_rejected, for example duplicate_payment or insufficient funds; the reason is also in the PAYMENT-REQUIRED header error field) or not settled (settlement_failed; the data is withheld).",
        "headers": {
          "PAYMENT-REQUIRED": {
            "$ref": "#/components/headers/PaymentRequired"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "payment_required",
                "message": "This route costs 0.01 USDC per call, paid with x402. Any x402 client (for example @x402/fetch with @x402/avm) pays it for you."
              },
              "docs": "/"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many failed paid attempts from this client in the last minute (rate_limited). Not charged.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          },
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Snapshot upstream is rate limiting (upstream_rate_limited) or the payment facilitator is unreachable (facilitator_unavailable). Not charged. Retry after the Retry-After seconds.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          },
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UpstreamTimeout": {
        "description": "Snapshot upstream timed out (upstream_timeout). Not charged.",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/CacheControlNoStore"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "upstream_timeout",
                "message": "Snapshot hub did not answer in time"
              }
            }
          }
        }
      }
    },
    "schemas": {
      "SpaceId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 64,
        "pattern": "^[^\\s/?#%\\\\]+\\.[^\\s/?#%\\\\]+$",
        "examples": ["ens.eth", "aavedao.eth"]
      },
      "ProposalId": {
        "type": "string",
        "pattern": "^(0x[0-9a-fA-F]{64}|Qm[1-9A-HJ-NP-Za-km-z]{44}|baf[a-z2-7]{56})$"
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "bad_request",
                  "not_found",
                  "payment_required",
                  "payment_rejected",
                  "settlement_failed",
                  "upstream_error",
                  "upstream_rate_limited",
                  "upstream_timeout",
                  "facilitator_error",
                  "facilitator_unavailable",
                  "method_not_allowed",
                  "misdirected_request",
                  "rate_limited",
                  "internal_error"
                ]
              },
              "message": {
                "type": "string"
              }
            }
          },
          "docs": {
            "type": "string",
            "description": "Where to read how to pay (402 only)."
          }
        }
      },
...
      "Space": {
        "type": "object",
        "required": ["id", "name", "network", "strategies", "stats", "link"],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": ["string", "null"]
          },
          "about": {
            "type": ["string", "null"]
          },
          "network": {
            "type": "string",
            "description": "Default chain id."
          },
          "symbol": {
            "type": ["string", "null"]
          },
          "avatar": {
            "type": ["string", "null"]
          },
          "website": {
            "type": ["string", "null"]
          },
          "twitter": {
            "type": ["string", "null"]
          },
          "github": {
            "type": ["string", "null"]
          },
          "coingecko": {
            "type": ["string", "null"]
          },
          "verified": {
            "type": "boolean"
          },
          "flagged": {
            "type": "boolean"
          },
          "hibernated": {
            "type": "boolean"
          },
          "turbo": {
            "type": "boolean"
          },
          "admins": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "moderators": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "membersCount": {
            "type": "integer"
          },
          "strategies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StrategyRef"
            }
          },
          "proposalValidation": {
            "type": ["string", "null"]
          },
          "voteValidation": {
            "type": ["string", "null"]
          },
          "voting": {
            "type": "object",
            "properties": {
              "delay": {
                "type": ["integer", "null"]
              },
              "period": {
                "type": ["integer", "null"]
              },
              "type": {
                "type": ["string", "null"]
              },
              "quorum": {
                "type": ["number", "null"]
              },
              "quorumType": {
                "type": ["string", "null"]
              },
              "blind": {
                "type": ["boolean", "null"]
              },
              "privacy": {
                "type": ["string", "null"]
              }
            }
          },
          "stats": {
            "type": "object",
            "properties": {
              "activeProposals": {
                "type": ["integer", "null"]
              },
              "proposals": {
                "type": ["integer", "null"]
              },
              "proposals30d": {
                "type": ["integer", "null"]
              },
              "followers": {
                "type": ["integer", "null"]
              },
              "votes": {
                "type": ["integer", "null"]
              },
              "votes7d": {
                "type": ["integer", "null"]
              }
            }
          },
          "treasuries": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": ["string", "null"]
                },
                "address": {
                  "type": ["string", "null"]
                },
                "network": {
                  "type": ["string", "null"]
                }
              }
            }
          },
          "delegationPortal": {
            "type": ["object", "null"],
            "properties": {
              "delegationType": {
                "type": "string"
              },
              "delegationContract": {
                "type": "string"
              },
              "delegationNetwork": {
                "type": "string"
              }
            }
          },
          "parent": {
            "type": ["string", "null"]
          },
          "children": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created": {
            "type": "integer"
          },
          "link": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
      "ProposalList": {
        "type": "object",
        "required": [
          "space",
          "state",
          "first",
          "skip",
          "hasMore",
          "next",
          "items"
        ],
        "properties": {
          "space": {
            "type": "string"
          },
          "state": {
            "type": ["string", "null"]
          },
          "first": {
            "type": "integer"
          },
          "skip": {
            "type": "integer"
          },
          "hasMore": {
            "type": "boolean",
            "description": "True when another page may exist."
          },
          "next": {
            "type": ["string", "null"],
            "description": "Relative URL of the next page, or null (last page, or past skip 5000)."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProposalSummary"
            }
          }
        }
      },
...
      "VoteList": {
        "type": "object",
        "required": [
          "proposal",
          "orderBy",
          "first",
          "skip",
          "total",
          "hasMore",
          "next",
          "items"
        ],
        "properties": {
          "proposal": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "space": {
                "type": ["string", "null"]
              },
              "state": {
                "type": "string"
              },
              "type": {
                "type": ["string", "null"]
              },
              "privacy": {
                "type": ["string", "null"]
              },
              "choices": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": ["id", "space", "state", "type", "privacy", "choices"]
          },
          "orderBy": {
            "type": "string",
            "enum": ["vp", "created"]
          },
          "first": {
            "type": "integer"
          },
          "skip": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "description": "Number of votes on the proposal."
          },
          "hasMore": {
            "type": "boolean",
            "description": "True when another page may exist."
          },
          "next": {
            "type": ["string", "null"],
            "description": "Relative URL of the next page, or null (last page, or past skip 5000)."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vote"
            }
          }
        }
      },
...
      "SettleResponse": {
        "type": "object",
        "required": ["success", "transaction", "network"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "transaction": {
            "type": "string",
            "description": "Algorand transaction id."
          },
          "network": {
            "type": "string"
          },
          "payer": {
            "type": "string"
          },
          "errorReason": {
            "type": "string"
          }
        }
      }
    }
  }
}